+++ CHANGELOG.md
@@ -10,8 +10,10 @@
     * **Note** Add retry logic for retryable HTTP status codes. [#147](https://github.com/patrickcping/pingone-go-sdk-v2/pull/147)
 * `github.com/patrickcping/pingone-go-sdk-v2/mfa` : [v0.9.2](./mfa/CHANGELOG.md)
     * **Note** Add retry logic for retryable HTTP status codes. [#147](https://github.com/patrickcping/pingone-go-sdk-v2/pull/147)
-* `github.com/patrickcping/pingone-go-sdk-v2/risk` : [v0.4.1](./risk/CHANGELOG.md)
+* `github.com/patrickcping/pingone-go-sdk-v2/risk` : [v0.5.0](./risk/CHANGELOG.md)
+    * **Breaking change** `RiskPredictor` model changed to `RiskPredictorCommon`, replaced with the predictor type model. [#151](https://github.com/patrickcping/pingone-go-sdk-v2/pull/151)
     * **Note** Add retry logic for retryable HTTP status codes. [#147](https://github.com/patrickcping/pingone-go-sdk-v2/pull/147)
+    * **Bug fix** Fix for Risk Predictor API request/responses not including the individual predictor type data models. [#151](https://github.com/patrickcping/pingone-go-sdk-v2/pull/151)
 
 # Release (2023-04-18)